requirements.txt
umap-learn>=0.3.7
numpy>=1.13
scikit-learn>=0.18
matplotlib>=3.0.3
